+++ sdk/textanalytics/Azure.AI.TextAnalytics/tests/ExtractSummaryTests.cs
@@ -5,6 +5,7 @@ using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
+using Azure.AI.TextAnalytics.Tests.Infrastructure;
 using Azure.Core.TestFramework;
 using NUnit.Framework;
 
@@ -78,6 +79,7 @@ namespace Azure.AI.TextAnalytics.Tests
         private const int ExtractSummaryMaxSentenceCount = 5;
 
         [RecordedTest]
+        [RetryOnInternalServerError]
         public async Task ExtractSummaryWithAADTest()
         {
             TextAnalyticsClient client = GetClient(useTokenCredential: true);
@@ -95,6 +97,7 @@ namespace Azure.AI.TextAnalytics.Tests
         }
 
         [RecordedTest]
+        [RetryOnInternalServerError]
         public async Task ExtractSummaryBatchWithRankOrderTest()
         {
             TextAnalyticsClient client = GetClient();
@@ -118,6 +121,7 @@ namespace Azure.AI.TextAnalytics.Tests
         }
 
         [RecordedTest]
+        [RetryOnInternalServerError]
         public async Task ExtractSummaryBatchWithErrorTest()
         {
             TextAnalyticsClient client = GetClient();
@@ -143,6 +147,7 @@ namespace Azure.AI.TextAnalytics.Tests
         }
 
         [RecordedTest]
+        [RetryOnInternalServerError]
         public async Task ExtractSummaryBatchConvenienceTest()
         {
             TextAnalyticsClient client = GetClient();
@@ -160,6 +165,7 @@ namespace Azure.AI.TextAnalytics.Tests
         }
 
         [RecordedTest]
+        [RetryOnInternalServerError]
         public async Task ExtractSummaryBatchConvenienceWithStatisticsTest()
         {
             TextAnalyticsClient client = GetClient();
@@ -183,6 +189,7 @@ namespace Azure.AI.TextAnalytics.Tests
         }
 
         [RecordedTest]
+        [RetryOnInternalServerError]
         public async Task ExtractSummaryBatchTest()
         {
             TextAnalyticsClient client = GetClient();
@@ -200,6 +207,7 @@ namespace Azure.AI.TextAnalytics.Tests
         }
 
         [RecordedTest]
+        [RetryOnInternalServerError]
         public async Task ExtractSummaryBatchWithStatisticsTest()
         {
             TextAnalyticsClient client = GetClient();
@@ -223,13 +231,12 @@ namespace Azure.AI.TextAnalytics.Tests
         }
 
         [RecordedTest]
-        [Ignore("https://github.com/Azure/azure-sdk-for-net/issues/32759")]
+        [RetryOnInternalServerError]
         public async Task ExtractSummaryBatchConvenienceWithAutoDetectedLanguageTest()
         {
             TextAnalyticsClient client = GetClient();
-            ExtractSummaryOptions options = new() { AutoDetectionDefaultLanguage = "en" };
 
-            ExtractSummaryOperation operation = await client.StartExtractSummaryAsync(s_extractSummaryBatchConvenienceDocuments, "auto", options);
+            ExtractSummaryOperation operation = await client.StartExtractSummaryAsync(s_extractSummaryBatchConvenienceDocuments, "auto");
             await operation.WaitForCompletionAsync();
             ValidateOperationProperties(operation);
 
@@ -242,7 +249,7 @@ namespace Azure.AI.TextAnalytics.Tests
         }
 
         [RecordedTest]
-        [Ignore("https://github.com/Azure/azure-sdk-for-net/issues/32759")]
+        [RetryOnInternalServerError]
         public async Task AnalyzeOperationExtractSummaryWithAutoDetectedLanguageTest()
         {
             TextAnalyticsClient client = GetClient();